[Utils] General Cleaning, strings file is deleted, jsoni file has been renamed to io (input & output)
